Tests/LibWeb/Text/expected/content-blocker-shadow-has-state-invalidation.txt

hover before: block
hover after enter: none
hover after leave: block
modal before: block
modal after showModal: none
modal after close: block
